练习sturt & currying

+++ exercises/basics/4.struct_test.go
@@ -6,10 +6,9 @@ import (
 )
 
 // 定义结构体
-
 type Person struct {
 	Name string
-	Age  int8
+	Age  int
 }
 
 // 接收值
